Create.html - added a container with a card housing a form with an input for petition title and submit button.
PetitionController - created a processCreate block of code with a post request method,  to deal with the input values from the create form. These values are added to the List petitions.
Updated the fragments.html to modify the navbar styling and piping in the branding location.
index.html added a thymeleaf for each method to list out the values submitted on the create page, to be developed further.