Create.html - added in description and author as inputs required from user
fragments.html - updated id's for future active state change
index.html - added in unless there is results - display the following message, added clickable links for results
Created the Petition class - title, description, author, and signatures - setters, getters, constructors and a toString
Updated the create post method to capture the values from the form elements, create a new petition object and populate it with the form data, add new petition to list of petitions
Added result mapping page, the mapping is passing in the title of the result
loop through the list to match the correct one
if found add to the model and return the result page.
Added result page - Populate a card with the title, description and author  and total amount of signatures - this part is a future consideration.